trying for a pwa.

+++ src/components/home.js
@@ -1,43 +1,61 @@
 import React, { useState, useEffect, createRef, Fragment } from "react";
 import { search, runIfProd } from "../utils";
+import { Img } from "./Img";
 
 function Home() {
   const [inputValue, setInputValue] = useState("");
   const [data, setData] = useState();
-  let imgRef = createRef();
+  let inputRef = createRef();
+  let lastQuery = localStorage.getItem("query");
 
   // Similar to componentDidMount and componentDidUpdate:
   useEffect(() => {
-    runIfProd(() => imgRef.current.focus());
-  });
+    runIfProd(() => inputRef.current.focus());
+    async function checkForLastQuery() {
+      if (lastQuery) {
+        setData(await search(lastQuery));
+      }
+    }
+    checkForLastQuery();
+    return () => {
+      lastQuery = "";
+    };
+  }, [inputValue, lastQuery]);
 
   return (
     <Fragment>
       <input
         value={inputValue}
-        placeholder={`Search for something`}
-        ref={imgRef}
+        placeholder={`Search for a gif`}
+        ref={inputRef}
         onChange={event => setInputValue(event.target.value)}
         onKeyPress={async event => {
           if (event.key === "Enter") {
             setData(await search(inputValue));
+            localStorage.setItem("query", inputValue);
           }
         }}
       />
       {data ? (
-        <section className="container">
+        <section className="gif-container">
           {" "}
           {data.map(img => (
-            <img
+            <Img
               src={img.webp}
-              alt={img.id}
+              alt={img.title}
               key={img.id}
               id={img.id}
               onClick={() => {
                 navigator.clipboard.writeText(img.url);
                 let data = JSON.parse(localStorage.getItem("data")) || [];
-                data.push(img.url);
-                localStorage.setItem("data", JSON.stringify(data));
+                let isAlreadyInStorage = data.filter(x => x.id === img.id)
+                  .length
+                  ? true
+                  : false;
+                if (!isAlreadyInStorage) {
+                  data.push(img);
+                  localStorage.setItem("data", JSON.stringify(data));
+                }
               }}
             />
           ))}
